+++ saleor/settings.py
@@ -509,6 +509,7 @@ PLUGINS_MANAGER = "saleor.plugins.manager.PluginsManager"
 
 PLUGINS = [
     "saleor.plugins.avatax.plugin.AvataxPlugin",
+    "saleor.plugins.invoicing.plugin.InvoicingPlugin",
     "saleor.plugins.vatlayer.plugin.VatlayerPlugin",
     "saleor.plugins.webhook.plugin.WebhookPlugin",
     "saleor.payment.gateways.dummy.plugin.DummyGatewayPlugin",
